integration.records: move tests to separate module
