+++ tests/integration/records/test_records_receivers.py
@@ -26,6 +26,7 @@ import os
 import pkg_resources
 import pytest
 import mock
+from elasticsearch import NotFoundError
 
 from invenio_db import db
 from invenio_oauthclient.utils import oauth_link_external_id
@@ -36,8 +37,10 @@ from invenio_oauthclient.models import (
     UserIdentity,
 )
 
-from inspirehep.modules.records.api import InspireRecord
 from inspirehep.modules.migrator.tasks import migrate_and_insert_record
+from inspirehep.modules.records.api import InspireRecord
+from inspirehep.modules.search import LiteratureSearch
+from inspirehep.utils.record import get_title
 
 from utils import _delete_record
 
@@ -274,3 +277,39 @@ def test_orcid_push_triggered_on_create_record_with_multiple_authors_with_allow_
     mocked_Task.apply_async.assert_any_call(**expected_kwargs_user1)
     mocked_Task.apply_async.assert_any_call(**expected_kwargs_user2)
     assert mocked_Task.apply_async.call_count == 2
+
+
+def test_that_db_changes_are_mirrored_in_es(app):
+    search = LiteratureSearch()
+    json = {
+        '$schema': 'http://localhost:5000/schemas/records/hep.json',
+        'document_type': [
+            'article',
+        ],
+        'titles': [
+            {'title': 'foo'},
+        ],
+        '_collections': ['Literature']
+    }
+
+    # When a record is created in the DB, it is also created in ES.
+
+    record = InspireRecord.create(json)
+    es_record = search.get_source(record.id)
+
+    assert get_title(es_record) == 'foo'
+
+    # When a record is updated in the DB, is is also updated in ES.
+
+    record['titles'][0]['title'] = 'bar'
+    record.commit()
+    es_record = search.get_source(record.id)
+
+    assert get_title(es_record) == 'bar'
+
+    # When a record is deleted in the DB, it is also deleted in ES.
+
+    record._delete(force=True)
+
+    with pytest.raises(NotFoundError):
+        es_record = search.get_source(record.id)
